# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WIZIA API Documentation",
        "description": "Documentation de l'API WIZIA",
        "contact": {
            "email": "contact@wizia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/ia/generateIALocal": {
            "post": {
                "tags": [
                    "Intelligence artificielle"
                ],
                "summary": "Génère une réponse texte avec Llama3.2",
                "operationId": "cee9ced13d67eabfd48c10d106f9401f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prompt"
                                ],
                                "properties": {
                                    "prompt": {
                                        "type": "string",
                                        "example": "Explique-moi Laravel en quelques mots"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réponse générée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "text": {
                                            "type": "string",
                                            "example": "Laravel est un framework PHP..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la génération",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur de la génération du prompt"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ia/generateIApicture": {
            "post": {
                "tags": [
                    "Intelligence artificielle"
                ],
                "summary": "Génère une image avec DALL-E 3",
                "operationId": "48b32bb083d72893714cea1e2a5dcb37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prompt"
                                ],
                                "properties": {
                                    "prompt": {
                                        "type": "string",
                                        "example": "Un chat astronaute dans l'espace"
                                    },
                                    "size": {
                                        "description": "Tailles disponibles: 256x256, 512x512, 1024x1024, 1024x1792",
                                        "type": "string",
                                        "example": "1024x1024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image générée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "image_url": {
                                            "type": "string",
                                            "example": "https://oaidalleapiprodscus.blob.core.windows.net/..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la génération de l'image",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Image non générée"
                                        },
                                        "details": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ia/generateIA": {
            "post": {
                "tags": [
                    "Intelligence artificielle"
                ],
                "summary": "Génère un post structuré pour les réseaux sociaux avec Gemini",
                "operationId": "e7556624ec966a71dede40674bda3324",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prompt"
                                ],
                                "properties": {
                                    "prompt": {
                                        "type": "string",
                                        "example": "Crée un post Instagram sur le café artisanal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post généré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Le café artisanal : un art à découvrir"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Découvrez les secrets du café de spécialité..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The prompt field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la génération",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur de la génération du prompt Gemini"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}